+++ src/features/pmicControl/npm/batteryProfiler.ts
@@ -13,6 +13,7 @@ import {
     LoggingEvent,
     Profile,
     ProfilingEvent,
+    ProfilingEventData,
     ProfilingState,
 } from './types';
 
@@ -21,7 +22,7 @@ export const BatteryProfiler: IBatteryProfiler = (
     eventEmitter: EventEmitter
 ) => {
     let profiling: ProfilingState = 'Off';
-    const processModuleCcProfiling = ({ message }: LoggingEvent) => {
+    const processModuleCcProfiling = ({ timestamp, message }: LoggingEvent) => {
         if (message.includes('Success: Profiling sequence completed')) {
             profiling = 'Ready';
             eventEmitter.emit('onProfilingStateChange', profiling);
@@ -35,7 +36,7 @@ export const BatteryProfiler: IBatteryProfiler = (
             eventEmitter.emit('onProfilingStateChange', profiling);
         } else {
             const messageParts = message.split(',');
-            const event: ProfilingEvent = {
+            const data: ProfilingEventData = {
                 iLoad: 0,
                 vLoad: 0,
                 tBat: 0,
@@ -50,35 +51,36 @@ export const BatteryProfiler: IBatteryProfiler = (
                 const pair = part.split('=');
                 switch (pair[0]) {
                     case 'iload':
-                        event.iLoad = Number.parseFloat(pair[1]);
+                        data.iLoad = Number.parseFloat(pair[1]);
                         break;
                     case 'vload':
-                        event.vLoad = Number.parseFloat(pair[1]);
+                        data.vLoad = Number.parseFloat(pair[1]);
                         break;
                     case 'tbat':
-                        event.tBat = Number.parseFloat(pair[1]);
+                        data.tBat = Number.parseFloat(pair[1]);
                         break;
                     case 'cycle':
-                        event.cycle = Number.parseInt(pair[1], 10);
+                        data.cycle = Number.parseInt(pair[1], 10);
                         break;
                     case 'seq':
-                        event.seq = Number.parseInt(pair[1], 10);
+                        data.seq = Number.parseInt(pair[1], 10);
                         break;
                     case 'chg':
-                        event.chg = Number.parseFloat(pair[1]);
+                        data.chg = Number.parseFloat(pair[1]);
                         break;
                     case 'rep':
-                        event.rep = Number.parseInt(pair[1], 10);
+                        data.rep = Number.parseInt(pair[1], 10);
                         break;
                     case 't0':
-                        event.t0 = Number.parseFloat(pair[1]);
+                        data.t0 = Number.parseFloat(pair[1]);
                         break;
                     case 't1':
-                        event.t1 = Number.parseFloat(pair[1]);
+                        data.t1 = Number.parseFloat(pair[1]);
                         break;
                 }
             });
 
+            const event: ProfilingEvent = { timestamp, data };
             eventEmitter.emit('onProfilingEvent', event);
         }
     };